Fix favorite marking functionality
Unfortunately the Status returned by the twitter api call doesn't always
reflect the change, even if it's taken place fine on the server.  This means I
can't rely on updating the list with the status to reflect it's update.

I've worked around this with an in memory list of favorites marked for this
session.

Without caching the list of favorites the star won't show on restarting the
app.  Both the official Twitter client and Plume seem to have the exact same
functionality.

I could cache the list of favorites in the future, but for simplicity will move
on for now.